requirements.txt
# Inventario PRO - Dependencias
kivy==2.3.0
openpyxl>=3.1.0
imaplib2>=3.6
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
fpdf2>=2.8.0
